Fixed bug, made new checking algorithm work 100%
The problem was that height and width stayed at 10, even when you changed them with
the boxes, so when you made the size bigger, it still stopped checking after
row 9 and column 9, but I made it so that height and width update with
the actual height and width. It's a lot faster, just as I'd suspected.