Readded handles, set up for dragging functionality.
I made part of the functionality for dragging the handle on angle c, I just
need to draw the lines while you drag, re-add the number labels, and make it
fill in boxes when you let go, and find the angle, and do the other 2 angles.
And then fix all the bugs that come up while I do that stuff.